requirements.txt
mpu6050-raspberrypi
RPi.GPIO
posix-ipc
evdev
